移除RunSection
