+++ src/section_run_server.py
@@ -1,9 +1,13 @@
+from enum import Enum
+import logging
 import os
 import json
 import math
+import subprocess
 from PySide6.QtCore import Qt
 from PySide6.QtWidgets import QVBoxLayout, QHBoxLayout, QLabel, QGroupBox
 from qfluentwidgets import (
+    ComboBox,
     PushButton,
     SpinBox,
     PrimaryPushButton,
@@ -14,13 +18,95 @@ from qfluentwidgets import (
     Slider,
 )
 
-from .common import CURRENT_DIR, CONFIG_FILE, RunSection
+from .common import CURRENT_DIR, CONFIG_FILE
 from .ui import *
-from .section_settings import SettingsSection  # 新增导入
 
-class RunServerSection(RunSection):
+
+class GPUType(Enum):
+    NVIDIA = 1
+    AMD = 2
+    UNKNOWN = 3
+
+
+class GPUManager:
+    def __init__(self):
+        self.nvidia_gpus = []
+        self.amd_gpus = []
+        self.detect_gpus()
+
+    def detect_gpus(self):
+        # 检测NVIDIA GPU
+        try:
+            os.environ["CUDA_DEVICE_ORDER"] = "PCI_BUS_ID"
+            result = subprocess.run(
+                "nvidia-smi --query-gpu=name --format=csv,noheader",
+                shell=True,
+                capture_output=True,
+                text=True,
+            )
+            if result.returncode == 0:
+                self.nvidia_gpus = result.stdout.strip().split("\n")
+        except Exception as e:
+            logging.error(f"检测NVIDIA GPU时出错: {str(e)}")
+
+        # 检测AMD GPU
+        try:
+            import wmi
+
+            c = wmi.WMI()
+            amd_gpus_temp = []
+            for gpu in c.Win32_VideoController():
+                if "AMD" in gpu.Name or "ATI" in gpu.Name:
+                    amd_gpus_temp.append(gpu.Name)
+            logging.info(f"检测到AMD GPU(正向列表): {amd_gpus_temp}")
+            # 反向添加AMD GPU
+            self.amd_gpus = list(reversed(amd_gpus_temp))
+            logging.info(f"检测到AMD GPU(反向列表): {self.amd_gpus}")
+        except Exception as e:
+            logging.error(f"检测AMD GPU时出错: {str(e)}")
+
+    def get_gpu_type(self, gpu_name):
+        if "NVIDIA" in gpu_name.upper():
+            return GPUType.NVIDIA
+        elif "AMD" in gpu_name.upper() or "ATI" in gpu_name.upper():
+            return GPUType.AMD
+        else:
+            return GPUType.UNKNOWN
+
+    def set_gpu_env(self, env, selected_gpu, selected_index, manual_index=None):
+        gpu_type = self.get_gpu_type(selected_gpu)
+        if manual_index == "":
+            manual_index = None
+        if gpu_type == GPUType.NVIDIA:
+            if manual_index is not None:
+                env["CUDA_VISIBLE_DEVICES"] = str(manual_index)
+                logging.info(f"设置 CUDA_VISIBLE_DEVICES = {manual_index}")
+            else:
+                env["CUDA_VISIBLE_DEVICES"] = str(selected_index)
+                logging.info(f"设置 CUDA_VISIBLE_DEVICES = {selected_index}")
+        elif gpu_type == GPUType.AMD:
+            if manual_index is not None:
+                env["HIP_VISIBLE_DEVICES"] = str(manual_index - len(self.nvidia_gpus))
+                logging.info(
+                    f"设置 HIP_VISIBLE_DEVICES = {manual_index - len(self.nvidia_gpus)}"
+                )
+            else:
+                env["HIP_VISIBLE_DEVICES"] = str(selected_index - len(self.nvidia_gpus))
+                logging.info(
+                    f"设置 HIP_VISIBLE_DEVICES = {selected_index - len(self.nvidia_gpus)}"
+                )
+        else:
+            logging.warning(f"未知的GPU类型: {selected_gpu}")
+        return env
+
+
+class RunServerSection(QFrame):
     def __init__(self, title, main_window, parent=None):
-        super().__init__(title, main_window, parent)
+        super().__init__(parent)
+        self.main_window = main_window
+        self.setObjectName(title.replace(" ", "-"))
+        self.title = title
+
         self._init_ui()
         self.load_presets()
         self.refresh_models()
@@ -62,8 +148,6 @@ class RunServerSection(RunSection):
         self.run_button.setFixedSize(110, 30)
         buttons_layout.addWidget(self.run_button)
 
-
-
         buttons_group = QGroupBox("")
         buttons_group.setStyleSheet(
             """ QGroupBox {border: 0px solid darkgray; background-color: #202020; border-radius: 8px;}"""
@@ -76,7 +160,10 @@ class RunServerSection(RunSection):
 
         layout.addLayout(UiRow("上下文长度 -c", self._create_context_length_layout()))
         layout.addLayout(
-            UiRow("工作线程数量 -np", UiSlider(self, "n_parallel", 1, 1, 32, 1, spinbox_fixed_width=140))
+            UiRow(
+                "工作线程数量 -np",
+                UiSlider(self, "n_parallel", 1, 1, 32, 1, spinbox_fixed_width=140),
+            )
         )
 
         self.context_per_thread_label = QLabel(self)
@@ -113,7 +200,7 @@ class RunServerSection(RunSection):
         ip_port_log_layout = QHBoxLayout()
 
         ip_layout = QVBoxLayout()
-        self.host_input = self._create_editable_combo_box(["127.0.0.1", "0.0.0.0"])
+        self.host_input = UiEditableComboBox(self, ["127.0.0.1", "0.0.0.0"])
         ip_layout.addWidget(QLabel("主机地址 --host"))
         ip_layout.addWidget(self.host_input)
 
@@ -123,9 +210,7 @@ class RunServerSection(RunSection):
         host_layout.addWidget(self.port_input)
 
         log_layout = QVBoxLayout()
-        self.log_format_combo = self._create_editable_combo_box(
-            ["none", "text", "json"]
-        )
+        self.log_format_combo = UiEditableComboBox(self, ["none", "text", "json"])
         log_layout.addWidget(QLabel("日志格式 --log-format"))
         log_layout.addWidget(self.log_format_combo)
 
@@ -148,14 +233,14 @@ class RunServerSection(RunSection):
     def _init_advance_options(self, layout):
         layout_extra_options = QHBoxLayout()
         layout_extra_options.setContentsMargins(0, 0, 0, 0)  # 设置内部边距
-        
+
         self.flash_attention_check = UiCheckBox(self, "启用 Flash Attention -fa", True)
         layout_extra_options.addWidget(self.flash_attention_check)
         layout_extra_options.addStretch(1)
-        
+
         self.no_mmap_check = UiCheckBox(self, "启用 --no-mmap", True)
         layout_extra_options.addWidget(self.no_mmap_check)
-        
+
         layout.addLayout(layout_extra_options)
 
         layout.addLayout(UiRow("配置预设选择", self._create_preset_options()))
@@ -201,6 +286,84 @@ class RunServerSection(RunSection):
         self.context_length_input.valueChanged.connect(self.update_slider_from_input)
 
         return layout
+
+    def _create_model_selection_layout(self):
+        layout = QHBoxLayout()
+        self.model_path = EditableComboBox(self)
+        self.model_path.setPlaceholderText("请选择模型路径")
+        self.refresh_model_button = PushButton(FIF.SYNC, "刷新", self)
+        self.refresh_model_button.clicked.connect(self.refresh_models)
+        self.refresh_model_button.setFixedWidth(140)
+        layout.addWidget(self.model_path)
+        layout.addWidget(self.refresh_model_button)
+        return layout
+
+    def refresh_models(self):
+        self.model_path.clear()
+        models = []
+        search_paths = [CURRENT_DIR] + self.main_window.get_model_search_paths()
+        logging.debug(f"搜索路径: {search_paths}")
+        for path in search_paths:
+            logging.debug(f"正在搜索路径: {path}")
+            if os.path.exists(path):
+                logging.debug(f"路径存在: {path}")
+                if os.path.isdir(path):
+                    logging.debug(f"路径是目录: {path}")
+                    for root, dirs, files in os.walk(path):
+                        logging.debug(f"正在搜索子目录: {root}")
+                        logging.debug(f"文件列表: {files}")
+                        for f in files:
+                            if f.endswith(".gguf"):
+                                full_path = os.path.join(root, f)
+                                logging.debug(f"找到模型文件: {full_path}")
+                                models.append(full_path)
+                else:
+                    logging.debug(f"路径不是目录: {path}")
+            else:
+                logging.debug(f"路径不存在: {path}")
+
+        logging.debug(f"找到的模型文件: {models}")
+
+        # 从设置中获取排序选项
+        sort_option = self.main_window.settings_section.model_sort_combo.currentText()
+
+        # 根据选择的排序方式对模型列表进行排序
+        if sort_option == "修改时间":
+            models.sort(key=lambda x: os.path.getmtime(x), reverse=True)
+        elif sort_option == "文件名":
+            models.sort(key=lambda x: os.path.basename(x).lower())
+        elif sort_option == "文件大小":
+            models.sort(key=lambda x: os.path.getsize(x), reverse=True)
+
+        self.model_path.addItems(models)
+
+    def _create_gpu_selection_layout(self):
+        layout = QHBoxLayout()
+        self.gpu_combo = ComboBox(self)
+        self.manully_select_gpu_index = LineEdit(self)
+        self.manully_select_gpu_index.setPlaceholderText("手动指定GPU索引")
+        self.manully_select_gpu_index.setFixedWidth(140)
+        layout.addWidget(self.gpu_combo)
+        layout.addWidget(self.manully_select_gpu_index)
+        return layout
+
+    def refresh_gpus(self):
+        self.gpu_combo.clear()
+        self.nvidia_gpus = self.main_window.gpu_manager.nvidia_gpus
+        self.amd_gpus = self.main_window.gpu_manager.amd_gpus
+
+        # 优先添加NVIDIA GPU
+        if self.nvidia_gpus:
+            self.gpu_combo.addItems(self.nvidia_gpus)
+
+        # 如果有AMD GPU，添加到列表末尾
+        if self.amd_gpus:
+            self.gpu_combo.addItems(self.amd_gpus)
+
+        if not self.nvidia_gpus and not self.amd_gpus:
+            logging.warning("未检测到NVIDIA或AMD GPU")
+
+        self.gpu_combo.addItems(["自动"])
 
     def context_to_slider(self, context):
         min_value = math.log(256)
@@ -365,7 +528,10 @@ class RunServerSection(RunSection):
         try:
             with open(config_file_path, "r", encoding="utf-8") as f:
                 config = json.load(f)
-            if config.get("remember_advanced_state", False) and self.main_window.settings_section.remember_advanced_state.isChecked():
+            if (
+                config.get("remember_advanced_state", False)
+                and self.main_window.settings_section.remember_advanced_state.isChecked()
+            ):
                 self.menu_advance.setVisible(config.get("advanced_state", False))
         except (FileNotFoundError, json.JSONDecodeError):
             pass
