+++ src/ui.py
@@ -1,21 +1,34 @@
 from PySide6.QtCore import Qt, QTimer, QSize
 from PySide6.QtWidgets import QWidget, QHBoxLayout, QVBoxLayout, QLabel, QFrame
-from qfluentwidgets import CheckBox, FluentStyleSheet, LineEdit, Slider, SpinBox
+from qfluentwidgets import (
+    CheckBox,
+    FluentStyleSheet,
+    LineEdit,
+    Slider,
+    SpinBox,
+    EditableComboBox,
+)
 from qfluentwidgets.common.style_sheet import (
-    StyleSheetManager, getStyleSheet, StyleSheetCompose, CustomStyleSheet
+    StyleSheetManager,
+    getStyleSheet,
+    StyleSheetCompose,
+    CustomStyleSheet,
 )
 from qfluentwidgets.common.config import qconfig
 
+
 class CustomStyleSheetManager(StyleSheetManager):
     def addCustomStyle(self, widget: QWidget, customStyle: str):
-        """ 添加自定义样式到小部件，而不覆盖现有样式 """
+        """添加自定义样式到小部件，而不覆盖现有样式"""
         if widget in self.widgets:
             source = self.widgets[widget]
             if isinstance(source, StyleSheetCompose):
-                custom_source = next((s for s in source.sources if isinstance(s, CustomStyleSheet)), None)
+                custom_source = next(
+                    (s for s in source.sources if isinstance(s, CustomStyleSheet)), None
+                )
                 if custom_source:
                     existing_style = custom_source.content()
-                    new_style = existing_style + '\n' + customStyle
+                    new_style = existing_style + "\n" + customStyle
                     custom_source.setCustomStyleSheet(new_style, new_style)
                 else:
                     custom_source = CustomStyleSheet(widget)
@@ -33,32 +46,39 @@ class CustomStyleSheetManager(StyleSheetManager):
         self.updateWidgetStyleSheet(widget)
 
     def updateWidgetStyleSheet(self, widget: QWidget):
-        """ 更新特定小部件的样式表 """
+        """更新特定小部件的样式表"""
         if widget in self.widgets:
             source = self.widgets[widget]
             widget.setStyleSheet(getStyleSheet(source, qconfig.theme))
 
+
 # 创建自定义样式表管理器的实例
 custom_style_manager = CustomStyleSheetManager()
 
+
 def addCustomWidgetStyle(widget: QWidget, customStyle: str):
-    """ 添加自定义样式到小部件，而不覆盖现有样式 """
+    """添加自定义样式到小部件，而不覆盖现有样式"""
     custom_style_manager.addCustomStyle(widget, customStyle)
+
 
 def UiCheckBox(parent, text, checked):
     w = CheckBox(parent)
     w.setText(text)
     w.setChecked(checked)
-    
+
     # 注册默认样式
     custom_style_manager.register(FluentStyleSheet.CHECK_BOX, w)
-    
+
     # 设置 CheckBox 的大小
     checkbox_size = 12  # 可以根据需要调整这个值
     w.setIconSize(QSize(checkbox_size, checkbox_size))
-    
+
     # 添加自定义样式（在下一个事件循环中应用，以确保覆盖其他样式）
-    QTimer.singleShot(0, lambda: addCustomWidgetStyle(w, f"""
+    QTimer.singleShot(
+        0,
+        lambda: addCustomWidgetStyle(
+            w,
+            f"""
         QCheckBox {{
             font-size: 12px !important;
         }}
@@ -66,8 +86,10 @@ def UiCheckBox(parent, text, checked):
             width: {checkbox_size}px;
             height: {checkbox_size}px;
         }}
-    """))
-    
+    """,
+        ),
+    )
+
     return w
 
 
@@ -122,14 +144,22 @@ def UiHLine(self):
     w.setFrameShadow(QFrame.Plain)  # 改为Plain以去除阴影效果
     w.setFixedHeight(32)
     # 使用更深的灰色作为背景色，并设置上下边距
-    w.setStyleSheet("""
+    w.setStyleSheet(
+        """
         background-color: #393939;
         margin-top: 15px;
         margin-bottom: 20px;
         max-height: 1px;
         border: none;  /* 去掉边框 */
-    """)
+    """
+    )
     return w
+
+
+def UiEditableComboBox(parent, items):
+    combo_box = EditableComboBox(parent)
+    combo_box.addItems(items)
+    return combo_box
 
 
 def UiRow(text, content):
@@ -140,6 +170,7 @@ def UiRow(text, content):
     else:
         layout.addLayout(content)
     return layout
+
 
 def UiCol(text, content):
     layout = QVBoxLayout()
